Refactor mobile_oauth.dart and request_code.dart
This commit refactors the mobile_oauth.dart and request_code.dart files. In mobile_oauth.dart, it adds an import statement for 'dart:io'. In request_code.dart, it adds an import statement for 'package:webview_windows/webview_windows.dart'. Additionally, it modifies the _performFullAuthFlow() method in mobile_oauth.dart to conditionally call the _requestCode.requestCode() method based on the platform. It also adds a new method, requestCodeWindows(), to request the code specifically for Windows platform. Finally, it updates the dependencies in pubspec.yaml to include the 'webview_windows' package.